Move container template in integration test to separate class
