+++ platform-tooling-support-tests/projects/jupiter-starter/src/test/resources/junit-platform.properties
@@ -1,0 +1,2 @@
+junit.jupiter.testclass.order.default = \
+  org.junit.jupiter.api.ClassOrderer$ClassName
